PottsGauge Documentation

PottsGauge.gauge — Method.

gauge(J,h,gauge::T)

Return the gauge-transform gauge of pair J,h. J,h are arrays of size q×q×N×N, q×N respectively. J should be symmetric: J[a,b,i,j] == J[b,a,j,i]. Allowed gauge are: ZeroSumGauge,LatticeGas,WildType, ExternalGauge.

Usage:

gauge(J,h,ZeroSumGauge())
gauge(J,h,Wildtype(x0))
gauge(J,h,ExternalGauge(U,V,C))

where:

* For `WildType` `x0` is a `Vector{Int}` of size `N` whose values are in the interval `1,…,q`

* For `ExternalGauge`, `U`,`V` are `q×N×N` arrays, and C is vector of length `N`.

source

PottsGauge.Energy — Method.

function Energy(J,h,x)

Compute the energy with parameters J::Array{T,4},h::Array{T,2} of a configuration x::Array{Int,1} of integer elements. x[i] ∈ 1:q where q is the number of letters (e.g. q=21 in MSA with gaps);

source

PottsGauge.testgauge — Method.

function testgauge(J1,h1,J2,h2; nsample::Integer)

Test if parameters J1,h1 and J2,h2 are gauge related; Return mean and std of the energy difference induced by the two set of parameters J1,h1 and J2,h2 of nsample random configurations.

source

PottsGauge Documentation

PottsGauge.gauge — Function.

gauge(J,h,gauge::T)

Return the gauge-transform gauge of pair J,h. J,h are arrays of size q×q×N×N, q×N respectively. J should be symmetric: J[a,b,i,j] == J[b,a,j,i]. Allowed gauge are: ZeroSumGauge,LatticeGas,WildType, ExternalGauge.

Usage:

gauge(J,h,ZeroSumGauge())
gauge(J,h,Wildtype(x0))
gauge(J,h,ExternalGauge(U,V,C))

where:

* For `WildType` `x0` is a `Vector{Int}` of size `N` whose values are in the interval `1,…,q`

* For `ExternalGauge`, `U`,`V` are `q×N×N` arrays, and C is vector of length `N`.

source