PottsGauge Documentation

PottsGauge.gauge — Method.

gauge(J,h,gauge::T)

Return the gauge-transform gauge of pair J,h. J,h are arrays of size q×q×N×N, q×N respectively. J should be symmetric: J[a,b,i,j] == J[b,a,j,i]. Allowed gauge are: ZeroSumGauge,LatticeGas,WildType.

Usage:

gauge(J,h,ZeroSumGauge())
gauge(J,h,Wildtype(x0))

where in the last case x0 is a Vector{Int} of size N whose values are in the interval 1,…,q

source

PottsGauge Documentation

PottsGauge.gauge — Function.

gauge(J,h,gauge::T)

Return the gauge-transform gauge of pair J,h. J,h are arrays of size q×q×N×N, q×N respectively. J should be symmetric: J[a,b,i,j] == J[b,a,j,i]. Allowed gauge are: ZeroSumGauge,LatticeGas,WildType.

Usage:

gauge(J,h,ZeroSumGauge())
gauge(J,h,Wildtype(x0))

where in the last case x0 is a Vector{Int} of size N whose values are in the interval 1,…,q

source